core/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
websockets>=12.0
pyyaml>=6.0
apscheduler>=3.10.0
openpyxl>=3.1.0
pydantic>=2.0.0
aiofiles>=23.0.0
jsonschema>=4.0.0
